// src/app/client/models/AssociatedContentResponse.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ContentResponse } from '../models/ContentResponse';
import { HttpFile } from '../http/http';

export class AssociatedContentResponse {
    'childs': Array<ContentResponse> | null;
    'parent': ContentResponse | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "childs",
            "baseName": "childs",
            "type": "Array<ContentResponse>",
            "format": ""
        },
        {
            "name": "parent",
            "baseName": "parent",
            "type": "ContentResponse",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AssociatedContentResponse.attributeTypeMap;
    }

    public constructor() {
    }
}